move search to alpha (#1902)
