+++ cmd/sealer/cmd/image/image.go
@@ -21,6 +21,6 @@ import (
 func NewImageCommands() []*cobra.Command {
 	var imageCommands []*cobra.Command
 	imageCommands = append(imageCommands, NewManifestCmd(), NewBuildCmd(), NewListCmd(), NewInspectCmd(), NewLoadCmd(),
-		NewLoginCmd(), NewLogoutCmd(), NewPullCmd(), NewPushCmd(), NewRmiCmd(), NewSaveCmd(), NewSearchCmd(), NewTagCmd())
+		NewLoginCmd(), NewLogoutCmd(), NewPullCmd(), NewPushCmd(), NewRmiCmd(), NewSaveCmd(), NewTagCmd())
 	return imageCommands
 }
